#200 Isolate en/learn and move getTableOfContentsFromCollection into tocUtils.ts
