+++ src/pages/contribute/index.astro
@@ -1,6 +1,6 @@
 ---
 import { getEntry, getCollection, type CollectionEntry } from "astro:content";
-import { getTableOfContentsFromCollection } from "~/utils/getTableOfContentsFromCollection";
+import { getTableOfContentsFromCollection } from "@utils/tocUtils";
 import SubCategoryLandingLayout from "@layouts/SubCategoryLandingLayout.astro";
 
 type SubCategoryDataType = Extract<
