+++ src/pages/en/[...slug].astro
@@ -1,42 +1,44 @@
 ---
-import { getTableOfContents } from "@utils/getTableOfContents";
+import { getSegmentToc } from "@utils/tocUtils";
 import { getFilteredDocEntries, getHeadings } from "@utils/contentUtils";
 import { getContentSlug } from "@utils/slugUtils";
 import MarkdownPage from "@components/MarkdownPage.astro";
 
 const language = "en" as const;
 
-//generating static paths from en collection
+// Generate static paths from EN collection excluding learn (implied) and specific segments.
 export async function getStaticPaths() {
   const language = "en" as const;
-  const excluded = ["automation", "database", "developer-portal", "mobile", "online", "onsite", "service", "ui"]; // Exclude specific segments
+  const excluded = ["automation", "database", "developer-portal", "mobile", "online", "onsite", "service", "ui"];
 
   // No redirects and not in excluded folders
   const docEntries = await getFilteredDocEntries(language, excluded);
   const headings = await getHeadings(docEntries);
 
-  return docEntries.map((entry, index) => {
-    const generatedSlug = getContentSlug(entry.filePath!, language);
+  const tocCache = new Map<string, Awaited<ReturnType<typeof getSegmentToc>>>();
 
-    const category = generatedSlug?.split("/")[0];
-    const isLearn = entry.data.uid?.startsWith("help");
+  return Promise.all(
+    docEntries.map(async (entry, i) => {
+      const slug = getContentSlug(entry.filePath!, language);
+      const segment = slug.split("/")[0];
 
-    function getTocPath(): string {
-      if (isLearn) return "learn";
-      return category;
-    }
+      if (!tocCache.has(segment)) {
+        tocCache.set(segment, await getSegmentToc(language, segment));
+      }
 
-    const tocPath = getTocPath();
-    const tocData = getTableOfContents(`docs/${language}/${tocPath}`);
-
-    return {
-      params: { slug: generatedSlug },
-      props: { entry, headings: headings[index], tocData, isLearn },
-    };
-  });
+      return {
+        params: { slug },
+        props: {
+          entry,
+          headings: headings[i],
+          tocData: tocCache.get(segment)!,
+        },
+      };
+    })
+  );
 }
 
-const { entry, headings, tocData, isLearn } = Astro.props;
+const { entry, headings, tocData } = Astro.props;
 ---
 
 <MarkdownPage
@@ -44,5 +46,4 @@ const { entry, headings, tocData, isLearn } = Astro.props;
   headings={headings}
   toc={tocData}
   language={language}
-  isLearn={isLearn}
 />
