+++ src/pages/en/learn/[...slug].astro
@@ -1,0 +1,37 @@
+---
+import { getHeadings, getDocEntries } from "@utils/contentUtils";
+import { getLocalizedToc } from "@utils/tocUtils";
+import { getContentSlug } from "@utils/slugUtils";
+import MarkdownPage from "@components/MarkdownPage.astro";
+
+const language = "en" as const;
+
+// Computed tocData once at module load (build-time) and reuse it for all pages
+const tocData = await getLocalizedToc(language);
+
+export async function getStaticPaths() {
+  const language = "en" as const;
+
+  const docEntries = await getDocEntries(language, "learn");
+  const headings = await getHeadings(docEntries);
+
+  return docEntries.map((entry, index) => {
+    const generatedSlug = getContentSlug(entry.filePath!, language);
+
+    return {
+      params: { slug: generatedSlug },
+      props: { entry, headings: headings[index] },
+    };
+  });
+}
+
+const { entry, headings } = Astro.props;
+---
+
+<MarkdownPage
+  entry={entry}
+  headings={headings}
+  toc={tocData}
+  language={language}
+  isLearn={true}
+/>
